+++ docs/release_notes.rst
@@ -4,7 +4,8 @@ Release Notes
 v0.1.4
 ------
 
-* Added `pages` in response for paginated responses
+* Added ``pages`` in response for paginated responses
+* Added Labels and Browser Application support
 
 v1.0.3
 ------